Remove kdtree spatial query from the CatalogueStore
The reason is that we don't expect the trees to be re-used many times, so the savings is lost to the construction time.  We also don't want the the memory use to increase over time as the trees are generated and saved.  Optimisation of the lookup should instead be done by the pixelization scheme.